requirements.txt
# Core ML & Data Science
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
scipy>=1.10.0

# Deep Learning
torch>=2.0.0
torchvision>=0.15.0
tensorflow>=2.13.0
transformers>=4.30.0

# XGBoost & Gradient Boosting
xgboost>=2.0.0
catboost>=1.2.0
lightgbm>=4.0.0

# Computer Vision
opencv-python>=4.8.0
Pillow>=10.0.0
albumentations>=1.3.0
ultralytics>=8.0.0  # YOLOv8

# Spectral Data Processing
spectral>=0.23.0
scikit-image>=0.21.0
peakutils>=1.3.4

# API & Web
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
pydantic>=2.0.0
requests>=2.31.0
aiohttp>=3.8.0

# Data Ingestion
beautifulsoup4>=4.12.0
feedparser>=6.0.10
selenium>=4.10.0
scrapy>=2.10.0

# Database & Storage
sqlalchemy>=2.0.0
pymongo>=4.4.0
redis>=4.6.0
h5py>=3.9.0

# Monitoring & Logging
prometheus-client>=0.17.0
mlflow>=2.5.0
wandb>=0.15.0
loguru>=0.7.0

# Explainability
shap>=0.42.0
lime>=0.2.0

# Dashboard
streamlit>=1.25.0
plotly>=5.15.0
dash>=2.11.0

# Workflow & Scheduling
prefect>=2.12.0
apache-airflow>=2.7.0

# Utilities
python-dotenv>=1.0.0
pyyaml>=6.0
tqdm>=4.65.0
joblib>=1.3.0
pyarrow>=12.0.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

# Data Annotation (optional)
# label-studio>=1.8.0

